Create my shell create child process
Write a C program to implement the shell which displays the command

prompt “myshell$”. It accepts the command, tokenize the command line and

execute it by creating the child process. Also implement the additional command

‘typeline’ as

typeline +n filename :- To print first n lines in the file.

typeline -a filename :- To print all lines in the file.