Create countcfilename.c
:Write a programto implement the toy shell. It should display the
command prompt “myshell$”. Tokenize the command line and execute the given command by
creating the child process. Additionally it should interpret the following commands.
count c filename :- To print number of characters in the file.

count w filename :- To print number of words in the file.
count l filename :- To print number of lines in the file.