Rename myshell$Listdirname.c to myshellListdirname.c
Write a program to implement the
shell. It should display the command prompt “myshell$”. Tokenize the command line and
execute the given command by creating the child process. Additionally it should interpret the
following ‘list’ commands as
myshell$ list f dirname:- To print names of all the files in current directory.
myshell$ list n dirname :- To print the number of all entries in the current directory